Tweak rollup.config.js
Create two separate configs, one for the browser that bundles all dependencies, and another for Node and bundlers that relies on dependencies being external.

For the umd build, tell plugin-node-resolve to prefer the `browser` property in package.json when bundling dependencies, which keeps us from packaging Node-specific code from isomorphic-unfetch.